# scripts/lib/api-zod/src/generated/types/adminDraw.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DrawSummary } from './drawSummary';

export type AdminDraw = DrawSummary & {
  draw_type: string;
  distributed: number;
  rollover: number;
  status: string;
};