# scripts/lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: string;
  name: string;
  email: string;
  plan: string;
  status: string;
  joined: string;
  scores: number;
}