# scripts/lib/api-zod/src/generated/types/adminWinner.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Winner } from './winner';

export type AdminWinner = Winner & ({
  user_name: string;
  user_email: string;
  /** @nullable */
  proof_url?: string | null;
});