# scripts/lib/api-zod/src/generated/types/charity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Charity {
  id: string;
  name: string;
  slug: string;
  short_description: string;
  full_description?: string;
  /** @nullable */
  logo_url?: string | null;
  /** @nullable */
  cover_image_url?: string | null;
  /** @nullable */
  website_url?: string | null;
  featured: boolean;
  active: boolean;
  impact_stat: string;
  focus?: string;
}