# scripts/lib/api-zod/src/generated/types/charityInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CharityInput {
  name: string;
  slug: string;
  short_description: string;
  full_description: string;
  focus: string;
  featured?: boolean;
}