# scripts/lib/api-zod/src/generated/types/charityPreferenceInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CharityPreferenceInput {
  charity_id: string;
  /**
     * @minimum 10
     * @maximum 100
     */
  contribution_percentage: number;
}