# scripts/lib/api-zod/src/generated/types/drawSimulationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DrawSimulationInputDrawType } from './drawSimulationInputDrawType';

export interface DrawSimulationInput {
  draw_type: DrawSimulationInputDrawType;
  /** @minimum 0 */
  prize_contribution: number;
  /** @minLength 1 */
  seed: string;
}