# scripts/lib/api-zod/src/generated/types/drawSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DrawSummary {
  id: string;
  label: string;
  status: string;
  draw_date: string;
  jackpot: number;
  entry_count: number;
  numbers?: number[];
}