# scripts/lib/api-zod/src/generated/types/prizeAllocation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PrizeAllocation {
  match_type: string;
  percentage: number;
  pool: number;
  winners: number;
  per_winner: number;
}