# scripts/lib/api-zod/src/generated/types/score.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Score {
  id: string;
  score: number;
  score_date: string;
  percentile: number;
  label: string;
}