# scripts/lib/api-zod/src/generated/types/scoreInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreInput {
  /**
     * @minimum 1
     * @maximum 45
     */
  score: number;
  score_date: Date;
}